requirements.txt
aiohttp
azure-cognitiveservices-speech
ffmpy
